+++ core/src/com/caletes/game/models/items/Item.java
@@ -5,13 +5,13 @@ import com.badlogic.gdx.graphics.g2d.Sprite;
 public abstract class Item {
 
     protected Sprite sprite;
-    private int originX, originY = 0;
+    private float originX, originY = 0;
 
     protected void setSprite(Sprite sprite) {
         this.sprite = sprite;
     }
 
-    protected void setOrigins(int originX, int originY) {
+    protected void setOrigins(float originX, float originY) {
         this.originX = originX;
         this.originY = originY;
     }
@@ -20,7 +20,7 @@ public abstract class Item {
         return sprite;
     }
 
-    public void setPosition(int x, int y) {
+    public void setPosition(float x, float y) {
         sprite.setPosition(x - originX, y - originY);
     }
 }
